requirements.txt
python-dotenv
pandas
numpy
matplotlib 
seaborn
pymango
pymongo[srv]==3.6
certifi
scikit-learn
dill
pyyaml

#refer to setup.py
#-e .
